requirements.txt
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5

PyYAML~=6.0.2